Updates for posible use with iOS

+++ Sources/NotenikLib/prefs/AppPrefs.swift
@@ -9,7 +9,7 @@
 //  terms of the MIT License (https://opensource.org/licenses/MIT).
 //
 
-import Cocoa
+import Foundation
 
 import NotenikUtils
 
@@ -20,13 +20,12 @@ public class AppPrefs {
     public static let shared = AppPrefs()
     
     let fileManager = FileManager.default
-    var tempDir: URL
+    var tempDir: URL? = nil
     
     let defaults = UserDefaults.standard
     
     let launchingKey    = "app-launching"
     let quickDeletesKey = "quick-deletes"
-    let fontSizeKey     = "font-size"
     let tagsSelectKey   = "tags-to-select"
     let tagsSuppressKey = "tags-to-suppress"
     let parentRealmParentKey = "parent-realm-parent"
@@ -49,17 +48,8 @@ public class AppPrefs {
     
     var _qd: Bool = false
     
-    var _edFS:   Float   = 13.0
-    var _edFSCG: CGFloat = 13.0
-    
     var _prp = ""
     public var parentRealmPath = ""
-    
-    var userFont       = NSFont.userFont(ofSize: 13.0)
-    var fixedPitchFont = NSFont.userFixedPitchFont(ofSize: 13.0)
-    
-    var userFontAttrs       = [NSAttributedString.Key.font: NSFont.userFont(ofSize: 13.0)]
-    var fixedPitchFontAttrs = [NSAttributedString.Key.font: NSFont.userFixedPitchFont(ofSize: 13.0)]
     
     public var pickLists = ValuePickLists()
     
@@ -87,7 +77,9 @@ public class AppPrefs {
     /// Private initializer to enforce usage of the singleton instance
     private init() {
         
-        tempDir = fileManager.temporaryDirectory
+        if #available(iOS 10.0, *) {
+            tempDir = fileManager.temporaryDirectory
+        }
         
         // Retrieve and log info about the current app.
         if let infoDictionary = Bundle.main.infoDictionary {
@@ -122,8 +114,6 @@ public class AppPrefs {
     
     func resetDefaults() {
         confirmDeletes = true
-        resetEditFontSize()
-        deriveRelatedFontFields()
         parentRealmParent = ""
         useCount = 0
         favoritesColumns = 4
@@ -139,14 +129,6 @@ public class AppPrefs {
         _lastURL = defaults.url(forKey: lastURLKey)
         
         _qd = defaults.bool(forKey: quickDeletesKey)
-        
-        let defaultFontSize = defaults.float(forKey: fontSizeKey)
-        if defaultFontSize == 0.0 {
-            defaults.set(_edFS, forKey: fontSizeKey)
-        } else {
-            _edFS = defaultFontSize
-        }
-        deriveRelatedFontFields()
         
         let tsel = defaults.string(forKey: tagsSelectKey)
         if tsel != nil {
@@ -254,19 +236,6 @@ public class AppPrefs {
         }
     }
     
-    var editFontSize: Float {
-        get {
-            return _edFS
-        }
-        set {
-            if newValue > 8.0 && newValue < 24.0 {
-                _edFS = newValue
-                defaults.set(_edFS, forKey: fontSizeKey)
-                deriveRelatedFontFields()
-            }
-        }
-    }
-    
     var parentRealmParent: String {
         get {
             return _prp
@@ -291,65 +260,6 @@ public class AppPrefs {
                 defaults.set(_prp, forKey: parentRealmParentKey)
             }
         }
-    }
-    
-    /// Prepare all the font fields we might need, adjusted to the latest size selection
-    func deriveRelatedFontFields() {
-        
-        _edFSCG = CGFloat(_edFS)
-        
-        userFont = NSFont.userFont(ofSize: _edFSCG)
-        fixedPitchFont = NSFont.userFixedPitchFont(ofSize: _edFSCG)
-        
-        userFontAttrs = [NSAttributedString.Key.font: userFont]
-        fixedPitchFontAttrs = [NSAttributedString.Key.font: fixedPitchFont]
-    }
-    
-    public func increaseEditFontSize(by: Float) {
-        editFontSize += by
-    }
-    
-    public func decreaseEditFontSize(by: Float) {
-        editFontSize -= by
-    }
-    
-    public func resetEditFontSize() {
-        editFontSize = 13.0
-    }
-    
-    public func setRegularFont(object: NSObject) {
-        if userFont != nil {
-            if let cb = object as? NSComboBox {
-                cb.font = userFont
-            } else if let textView = object as? NSTextView {
-                textView.font = userFont
-            } else if let textField = object as? NSTextField {
-                textField.font = userFont
-            } else if let menu = object as? NSMenu {
-                menu.font = userFont
-            } 
-        }
-    }
-    
-    public func setFixedPitchFont(view: NSView) {
-        if userFont != nil {
-            if let textView = view as? NSTextView {
-                textView.font = fixedPitchFont
-            } else if let textField = view as? NSTextField {
-                textField.font = fixedPitchFont
-            }
-        }
-    }
-    
-    /// Make an attributed string using latest font size
-    public func makeUserAttributedString(text: String) -> NSAttributedString {
-        return NSAttributedString(string: text, attributes: userFontAttrs as [NSAttributedString.Key: Any])
-    }
-    
-    /// Determine the appropriate height constraint for a text view based on the desired
-    /// Number of lines to show. 
-    public func getViewHeight(lines: Float) -> CGFloat {
-        return CGFloat(editFontSize * 1.20 * lines)
     }
     
     public var tagsToSelect: String {
@@ -461,13 +371,14 @@ public class AppPrefs {
     }
     
     /// Given a file extension, return the next temp file to be used by Notenik.
-    func nextTempFile(ext: String) -> URL {
+    func nextTempFile(ext: String) -> URL? {
+        guard tempDir != nil else { return nil }
         var fnm = "notenik-temp-\(_tempFileCount)"
         _tempFileCount += 1
         if ext.count > 0 {
             fnm.append(".\(ext)")
         }
-        let tempFile = tempDir.appendingPathComponent(fnm)
+        let tempFile = tempDir!.appendingPathComponent(fnm)
         do {
             try fileManager.removeItem(at: tempFile)
         } catch {
